Add authentication middleware
